LayoutTests/http/tests/media/video-player-has-video-track.html

<!-- 
<html>

<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>

<body>
<script src="script-tests/get-network-services.js"></script>
<script src="../fast/js/resources/js-test-post.js"></script>
</body>

</html>
 -->


<html>

<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>

<body>
<script>
var vid;

function getInfo() {
	vid = document.getElementsByTagName("video")[0];
	shouldBe("vid.videoTracks.length>0", "true");
	shouldBe("vid.videoTracks.item(0).selected", "true");
	//window.console.log("getInfo():  video.videoTracks.len: " + vid.videoTracks.length);
	vid.parentNode.removeChild(vid);
	finishJSTest();
}

function insertCtrl() {
	var body = document.getElementsByTagName("body")[0];
	var v = document.createElement("video");
	v.setAttribute('controls', 'yes');
	var src = document.createElement("source");
	src.setAttribute('src', 'resources/sintel_trailer-480p.ogv');
	v.appendChild(src);
	body.appendChild(v);
}

insertCtrl();
description('Tests HTTPMediaElement Player.');
window.setTimeout('getInfo();',500);
window.jsTestIsAsync = true;
</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>

</html>
